requirements.txt
asgiref
Django
pytz==2021.3
sqlparse
django-ckeditor
django-constance[database]
django-picklefield
sorl-thumbnail
Pillow
django-allauth
django-apscheduler
celery
redis
twilio
hiredis
ipython
flake8
channels
environs[django]
psycopg2-binary
gunicorn
daphne